+++ test/jest.config.cjs
@@ -2,7 +2,7 @@ module.exports = {
   preset: 'ts-jest',
   rootDir: '..',
   testEnvironment: 'node',
-  globals: { 'ts-jest': { tsconfig: '<rootDir>/test/tsconfig.json' } },
+  transform: { '^.+\\.tsx?$': ['ts-jest', { tsconfig: '<rootDir>/test/tsconfig.json' }] },
   roots: ['<rootDir>/test'],
   testMatch: ['**/*.spec.ts'],
   collectCoverageFrom: ['src/**/*.ts'],